Protect images in the integration/ suite
