+++ hack/make/test-integration-shell
@@ -5,3 +5,5 @@ bundle .integration-daemon-setup
 
 export ABS_DEST
 bash +e
+
+bundle .integration-daemon-stop
