+++ integration/container/create_test.go
@@ -2,42 +2,13 @@ package container
 
 import (
 	"context"
-	"fmt"
-	"os"
-	"testing"
-
 	"strconv"
+	"testing"
 
 	"github.com/docker/docker/api/types/container"
 	"github.com/docker/docker/api/types/network"
-	"github.com/docker/docker/client"
-	"github.com/docker/docker/integration-cli/environment"
-	"github.com/docker/docker/integration-cli/fixtures/load"
 	"github.com/stretchr/testify/require"
 )
-
-var (
-	testEnv *environment.Execution
-)
-
-func TestMain(m *testing.M) {
-	var err error
-	testEnv, err = environment.New()
-	if err != nil {
-		fmt.Println(err)
-		os.Exit(1)
-	}
-
-	if testEnv.LocalDaemon() {
-		fmt.Println("INFO: Testing against a local daemon")
-	} else {
-		fmt.Println("INFO: Testing against a remote daemon")
-	}
-
-	// TODO: ensure and protect images
-	res := m.Run()
-	os.Exit(res)
-}
 
 func TestAPICreateWithNotExistImage(t *testing.T) {
 	defer setupTest(t)()
@@ -117,25 +88,5 @@ func TestAPICreateEmptyEnv(t *testing.T) {
 			require.Error(t, err)
 			require.Contains(t, err.Error(), tc.expectedError)
 		})
-	}
-}
-
-func createClient(t *testing.T) client.APIClient {
-	clt, err := client.NewEnvClient()
-	require.NoError(t, err)
-	return clt
-}
-
-func setupTest(t *testing.T) func() {
-	if testEnv.DaemonPlatform() == "linux" {
-		images := []string{"busybox:latest", "hello-world:frozen", "debian:jessie"}
-		err := load.FrozenImagesLinux(testEnv.DockerBinary(), images...)
-		if err != nil {
-			t.Fatalf("%+v", err)
-		}
-		defer testEnv.ProtectImage(t, images...)
-	}
-	return func() {
-		testEnv.Clean(t, testEnv.DockerBinary())
 	}
 }
